requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic[email]==2.5.3

pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
# weasyprint==60.2 <-- Optional, frontend PDF generation used instead
jinja2==3.1.3
aiosmtplib==3.0.1
